// src/components/ui/layout-types.ts
export enum MainAxisAlignment {
  start = "justify-start",
  end = "justify-end",
  center = "justify-center",
  spaceBetween = "justify-between",
  spaceAround = "justify-around",
  spaceEvenly = "justify-evenly",
}

export enum CrossAxisAlignment {
  start = "items-start",
  end = "items-end",
  center = "items-center",
  stretch = "items-stretch",
  baseline = "items-baseline",
}

export enum MainAxisSize {
  min = "flex-none",
  max = "flex-1",
}

export enum BoxFit {
  fill = "object-fill",
  contain = "object-contain",
  cover = "object-cover",
  none = "object-none",
  scaleDown = "object-scale-down",
}

export enum Alignment {
  topLeft = "items-start justify-start",
  topCenter = "items-start justify-center",
  topRight = "items-start justify-end",
  centerLeft = "items-center justify-start",
  center = "items-center justify-center",
  centerRight = "items-center justify-end",
  bottomLeft = "items-end justify-start",
  bottomCenter = "items-end justify-center",
  bottomRight = "items-end justify-end",
}